refactor : extract server & socket function to separate files
