+++ tsconfig.prod.json
@@ -1,0 +1,4 @@
+{
+  "extends": "./tsconfig",
+  "exclude": ["**/*.test.ts", "**/*.mock.ts"]
+}
